@@ -3,41 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "serve -s dist",
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
+    "@azure/cosmos": "^4.0.0",
     "@azure/msal-browser": "^2.38.0",
     "core-js": "^3.6.4",
-    "vue": "^2.6.11"
+    "vue": "^3.3.4",
+    "vuetify": "^3.3.21"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.2.0",
-    "@vue/cli-plugin-eslint": "~4.2.0",
-    "@vue/cli-service": "~4.2.0",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.1.2",
-    "vue-template-compiler": "^2.6.11"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ]
+    "@mdi/font": "^7.3.67",
+    "@vitejs/plugin-vue": "^4.3.4",
+    "sass": "^1.69.3",
+    "sass-loader": "^13.3.2",
+    "vite": "^4.4.9"
+  }
 }
