@@ -1,8 +1,19 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 
-Vue.config.productionTip = false
+import '@mdi/font/css/materialdesignicons.css' // Ensure you are using css-loader
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+// Vuetify
+import 'vuetify/styles'
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
+const vuetify = createVuetify({
+  components,
+  directives,
+})
+
+
+createApp(App).use(vuetify).mount('#app')
+
